Splitting Graph API and GraphImpl
